oschannel实现优化，支持业务代态动态选择
