+++ packages/taro-platform-harmony-hybrid/src/api/apis/base/getAppAuthorizeSetting/index.ts
@@ -1,0 +1,8 @@
+import Taro from '@tarojs/api'
+
+import { getAppAuthorizeSetting as nativeGetAppAuthorizeSetting } from './native'
+import { getAppAuthorizeSetting as osChannelGetAppAuthorizeSetting } from './oschannel'
+
+export const getAppAuthorizeSetting: typeof Taro.getAppAuthorizeSetting = (useNativeImpl: boolean = false) => {
+  return useNativeImpl ? nativeGetAppAuthorizeSetting() : osChannelGetAppAuthorizeSetting()
+}
