+++ packages/taro-platform-harmony-hybrid/src/api/apis/base/getSystemSetting/index.ts
@@ -1,0 +1,8 @@
+import Taro from '@tarojs/api'
+
+import { getSystemSetting as nativeGetSystemSetting } from './native'
+import { getSystemSetting as osChannelGetSystemSetting } from './oschannel'
+
+export const getSystemSetting: typeof Taro.getSystemSetting = (useNativeImpl: boolean = false) => {
+  return useNativeImpl ? nativeGetSystemSetting() : osChannelGetSystemSetting()
+}
