+++ packages/taro-platform-harmony-hybrid/src/api/apis/base/system.ts
@@ -3,6 +3,8 @@ import Taro from '@tarojs/api'
 import native from '../NativeApi'
 import { shouldBeObject } from '../utils'
 import { MethodHandler } from '../utils/handler'
+import { getAppAuthorizeSetting } from './getAppAuthorizeSetting'
+import { getSystemSetting } from './getSystemSetting'
 
 /**
  * 跳转系统蓝牙设置页
@@ -80,30 +82,6 @@ export const getWindowInfo: typeof Taro.getWindowInfo = () => {
   return windowInfo
 }
 
-const lastSystemSettingResult: Taro.getSystemSetting.Result = {}
-let lastGetSystemSettingTime = 0
-
-/**
- * 获取设备设置
- *
- * @canUse getSystemSetting
- * @__return [bluetoothEnabled, locationEnabled, wifiEnabled, deviceOrientation[portrait, landscape]]
- */
-export const getSystemSetting: typeof Taro.getSystemSetting = () => {
-  const currentTime = Date.now()
-  if (currentTime - lastGetSystemSettingTime < 500) {
-    return lastSystemSettingResult
-  }
-  // @ts-ignore
-  const info = native.getSystemSetting()
-  lastSystemSettingResult.bluetoothEnabled = info.bluetoothEnabled
-  lastSystemSettingResult.locationEnabled = info.locationEnabled
-  lastSystemSettingResult.wifiEnabled = info.wifiEnabled
-  lastSystemSettingResult.deviceOrientation = info.deviceOrientation
-  lastGetSystemSettingTime = currentTime
-  return lastSystemSettingResult
-}
-
 /**
  * 获取设备基础信息
  *
@@ -151,37 +129,6 @@ export const getAppBaseInfo: typeof Taro.getAppBaseInfo = () => {
   return appBaseInfo
 }
 
-/**
- * 获取APP授权设置
- *
- * @canUse getAppAuthorizeSetting
- * @__return
- * [albumAuthorized, bluetoothAuthorized, cameraAuthorized, locationAuthorized, locationReducedAccuracy,\
- * microphoneAuthorized, notificationAuthorized, phoneCalendarAuthorized]
- */
-export const getAppAuthorizeSetting: typeof Taro.getAppAuthorizeSetting = () => {
-  const info = native.getAppAuthorizeSetting()
-  // @ts-ignore
-  const appAuthorizeSetting: Taro.getAppAuthorizeSetting.Result = {
-    /** 允许微信使用相册的开关（仅 iOS 有效） */
-    albumAuthorized: info.albumAuthorized,
-    /** 允许微信使用蓝牙的开关（仅 iOS 有效） */
-    bluetoothAuthorized: info.bluetoothAuthorized,
-    /** 允许微信使用摄像头的开关 */
-    cameraAuthorized: info.cameraAuthorized,
-    /** 允许微信使用定位的开关 */
-    locationAuthorized: info.locationAuthorized,
-    /** 定位准确度。true 表示模糊定位，false 表示精确定位（仅 iOS 有效） */
-    locationReducedAccuracy: info.locationAccuracy === 'reduced',
-    /** 允许微信使用麦克风的开关 */
-    microphoneAuthorized: info.microphoneAuthorized,
-    /** 允许微信通知的开关 */
-    notificationAuthorized: info.notificationAuthorized,
-    /** 允许微信读写日历的开关 */
-    phoneCalendarAuthorized: info.phoneCalendarAuthorized,
-  }
-  return appAuthorizeSetting
-}
 
 /**
  * 获取系统信息Taro.getSystemInfo的同步版本
