+++ packages/taro-platform-harmony-hybrid/src/api/apis/device/keyBoard/hideKeyboard/index.ts
@@ -1,0 +1,9 @@
+import Taro from '@tarojs/taro'
+
+import { hideKeyboard as nativeHideKeyboard } from './native'
+import { hideKeyboard as osChannelHideKeyboard } from './oschannel'
+
+
+export const hideKeyboard: typeof Taro.hideKeyboard = (options, useNativeImpl: boolean = false) => {
+  return useNativeImpl ? nativeHideKeyboard(options) : osChannelHideKeyboard(options)
+}
