+++ packages/taro-platform-harmony-hybrid/src/api/apis/device/phone/index.ts
@@ -1,1 +1,8 @@
-export * from './phone'
+import Taro from '@tarojs/api'
+
+import { makePhoneCall as nativeMakePhoneCall } from './native'
+import { makePhoneCall as osChannelMakePhoneCall } from './oschannel'
+
+export const makePhoneCall: typeof Taro.makePhoneCall = (options, useNativeImpl = false) => {
+  return useNativeImpl ? nativeMakePhoneCall(options) : osChannelMakePhoneCall(options)
+}
